Add unauthorized response type

+++ Shop.Api/Controllers/CartController.cs
@@ -25,10 +25,16 @@ namespace Shop.Api.Controllers
         }
 
         [HttpPost("{productId}")]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(CartItemDto))]
         public async Task<IActionResult> AddToCartAsync(Guid productId)
         {
             Request.Headers.TryGetValue(Authorization, out var token);
             var userData = await _identityApiService.GetUserData(token);
+            if (userData == null)
+            {
+                return Unauthorized();
+            }
 
             var updatedCartItem = _cartItemsService.AddToCart(productId, userData.Email);
             var cartItemDto = _mapper.Map<CartItemDto>(updatedCartItem);
@@ -37,10 +43,16 @@ namespace Shop.Api.Controllers
         }
 
         [HttpDelete("{productId}")]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<IActionResult> RemoveFromCartAsync(Guid productId)
         {
             Request.Headers.TryGetValue(Authorization, out var token);
             var userData = await _identityApiService.GetUserData(token);
+            if (userData == null)
+            {
+                return Unauthorized();
+            }
 
             _cartItemsService.RemoveFromCard(productId, userData.Email);
 
@@ -48,10 +60,16 @@ namespace Shop.Api.Controllers
         }
 
         [HttpPut("{productId}/reduce")]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<IActionResult> ReduceProductCountAsync(Guid productId)
         {
             Request.Headers.TryGetValue(Authorization, out var token);
             var userData = await _identityApiService.GetUserData(token);
+            if (userData == null)
+            {
+                return Unauthorized();
+            }
 
             _cartItemsService.ReduceProductCount(productId, userData.Email);
 
@@ -59,21 +77,33 @@ namespace Shop.Api.Controllers
         }
 
         [HttpGet]
-        public async Task<IEnumerable<CartItemDto>> GetCartItemsAsync()
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<CartItemDto>))]
+        public async Task<IActionResult> GetCartItemsAsync()
         {
             Request.Headers.TryGetValue(Authorization, out var token);
             var userData = await _identityApiService.GetUserData(token);
+            if (userData == null)
+            {
+                return Unauthorized();
+            }
 
             var items = _cartItemsService.GetCartItems(userData.Email);
 
-            return _mapper.Map<List<CartItemDto>>(items);
+            return Ok(_mapper.Map<List<CartItemDto>>(items));
         }
 
         [HttpGet("user-data")]
-        public async Task<UserDataDto> GetAllUserData()
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(UserDataDto))]
+        public async Task<IActionResult> GetAllUserData()
         {
             Request.Headers.TryGetValue(Authorization, out var token);
             var userData = await _identityApiService.GetUserData(token);
+            if (userData == null)
+            {
+                return Unauthorized();
+            }
 
             var result = new UserDataDto
             {
@@ -83,7 +113,7 @@ namespace Shop.Api.Controllers
                 TotalProductsPrice = _cartItemsService.GetTotalPrice(userData.Role),
             };
 
-            return result;
+            return Ok(result);
         }
     }
 }
